Move nodeSet into its own file
